Change the image to be based on UBI8
This is in preparation to have a multi-arch build for both Intel (x86_64) and
IBM Z (s390x). However not mandatory (we could base on RHEL7), it was put as
a desirable requirement - and might easy the process in some points.

Changes/fixes include:
- Change the base image
- Change the package manager to dnf
- Remove install of yum-utils
- Remove modules not available in UBI8 (including database drivers for
  postgresql and mysql - also remove corresponding datasource definitions)
- Change Maven installation method (from using Software Collections to using
  common packages)

References: IBMZP-28, JDG-3368

WARNING: the image might be in a non-buildable state after this commit,
  required changes might be present in future commits.